requirements.txt
openai
python-dotenv
Flask
gunicorn
flask_cors